requirements.txt
slack-sdk==3.21.3
textblob==0.17.1
vaderSentiment==3.3.2
pandas==2.0.3
matplotlib==3.7.2
plotly==5.15.0
python-dotenv==1.0.0
emoji==2.8.0
openai==1.3.0
flask==2.3.3
flask-socketio==5.3.6
python-socketio==5.9.0